Refactoring get_batch_results to be a generator
In an effort to provide more flexibility to the api client consumer,
get_batch_results now yields each result record as it is processed.
Because consumers may want not want to process result files serially,
this method is now limited to 1 result_id. A separate method has been
provided to get all of the result ids for a given batch.